Move setAll helpers up to the collection level
